add polarization support; result consistent with water_pol_1024

+++ admp/settings.py
@@ -2,7 +2,7 @@
 from jax import jit
 from jax.config import config
 
-PRECISION = 'single'  # 'double'
+PRECISION = 'double'  # 'double'
 
 DO_JIT = True
